# gmall-admin/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: guli-fast是一个轻量级的Java快速开发平台，能快速开发项目并交付【接私活利器】
  version: 1.0.0
  title: 谷粒快速开发平台
basePath: /guli-fast
schemes:
- http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/config/list:
    get:
      tags:
      - 参数管理
      summary: 参数列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: key
        description: 参数名
        in: query
        type: string
      responses:
        '200':
          description: 返回参数列表
          schema:
            $ref: '#/definitions/SysConfigEntityList'
  /sys/config/info/{id}:
    get:
      tags:
      - 参数管理
      summary: 获取参数信息
      produces:
      - application/json
      parameters:
      - name: id
        description: 参数ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回参数信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回参数信息
                $ref: '#/definitions/SysConfigEntity'
  /sys/config/save:
    post:
      tags:
      - 参数管理
      summary: 添加参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysConfigEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/update:
    post:
      tags:
      - 参数管理
      summary: 修改参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysConfigEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/delete:
    post:
      tags:
      - 参数管理
      summary: 删除参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/list:
    get:
      tags:
      - 文件服务
      summary: 文件列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      responses:
        '200':
          description: 返回文件列表
          schema:
            $ref: '#/definitions/SysOssEntityList'
...
  /sys/oss/saveConfig:
    post:
      tags:
      - 文件服务
      summary: 保存云存储配置信息
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysCloudStorageEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/upload:
    post:
      tags:
      - 文件服务
      summary: 上传文件
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - name: file
        description: 文件
        in: formData
        type: file
        required: true
      responses:
        '200':
          description: 返回文件列表
          schema:
            $ref: '#/definitions/FileUpload'
  /sys/oss/delete:
    post:
      tags:
      - 文件服务
      summary: 删除文件
      produces:
      - application/json
      parameters:
      - name: body
        description: 文件ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/list:
    get:
      tags:
      - 定时任务
      summary: 定时任务列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: beanName
        description: spring bean名称
        in: query
        type: string
      responses:
        '200':
          description: 返回定时任务列表
          schema:
            $ref: '#/definitions/ScheduleJobEntityList'
  /sys/schedule/info/{jobId}:
    get:
      tags:
      - 定时任务
      summary: 获取定时任务信息
      produces:
      - application/json
      parameters:
      - name: jobId
        description: 定时任务ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回定时任务信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              schedule:
                description: 定时任务信息
                $ref: '#/definitions/ScheduleJobEntity'
  /sys/schedule/save:
    post:
      tags:
      - 定时任务
      summary: 添加定时任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/ScheduleJobEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/update:
    post:
      tags:
      - 定时任务
      summary: 修改定时任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/ScheduleJobEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/delete:
    post:
      tags:
      - 定时任务
      summary: 删除定时任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/run:
    post:
      tags:
      - 定时任务
      summary: 立即执行任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/pause:
    post:
      tags:
      - 定时任务
      summary: 暂停定时任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/resume:
    post:
      tags:
      - 定时任务
      summary: 恢复定时任务
      produces:
      - application/json
      parameters:
      - name: body
        description: 定时任务ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  SysConfigEntity:
    type: object
    properties:
      id:
        description: 参数ID
        type: integer
        format: int64
      key:
        description: 参数名
        type: string
      value:
        description: 参数值
        type: string
      remark:
        description: 备注
        type: string
  SysConfigEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysConfigEntity'
  SysOssEntity:
    type: object
    properties:
      id:
        description: ID
        type: integer
        format: int64
      url:
        description: URL地址
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysOssEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysOssEntity'
  SysCloudStorageEntity:
    type: object
    properties:
      type:
        description: 类型 1：七牛  2：阿里云  3：腾讯云
        type: integer
        format: int32
      qiniuDomain:
        description: 七牛绑定的域名
        type: string
      qiniuPrefix:
        description: 七牛路径前缀
        type: string
      qiniuAccessKey:
        description: 七牛ACCESS_KEY
        type: string
      qiniuSecretKey:
        description: 七牛SECRET_KEY
        type: string
      qiniuBucketName:
        description: 七牛存储空间名
        type: string
      aliyunDomain:
        description: 阿里云绑定的域名
        type: string
      aliyunPrefix:
        description: 阿里云路径前缀
        type: string
      aliyunEndPoint:
        description: 阿里云EndPoint
        type: string
      aliyunAccessKeyId:
        description: 阿里云AccessKeyId
        type: string
      aliyunAccessKeySecret:
        description: 阿里云AccessKeySecret
        type: string
      aliyunBucketName:
        description: 阿里云BucketName
        type: string
      qcloudDomain:
        description: 腾讯云绑定的域名
        type: string
      qcloudPrefix:
        description: 腾讯云路径前缀
        type: string
      qcloudAppId:
        description: 腾讯云AppId
        type: string
      qcloudSecretId:
        description: 腾讯云SecretId
        type: string
      qcloudSecretKey:
        description: 腾讯云SecretKey
        type: string
      qcloudBucketName:
        description: 腾讯云BucketName
        type: string
      qcloudRegion:
        description: 腾讯云COS所属地区
        type: string
  FileUpload:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      url:
        description: 文件URL地址
        type: string
      msg:
        description: 失败原因
        type: string
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  ScheduleJobEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/ScheduleJobEntity'
...